build: modify app deployment to use nodeshift
This commit modifies the `package.json` file significantly in order to ease
deployment to an OpenShift environment by using `nodeshift`. To deploy to
an OpenShift instance now, you just need to be logged into the instance on
the command line, and then just run `npm run deploy`. This will build the
static site resources and use `nodeshift` to deploy the application, creating
the DeploymentConfig, BuildConfiguration, Service and Route.